crypto-plugin: split up test cases
Every backend gets its own test suite. This makes the identification of memory leaks easier.
