+++ src/plugins/crypto/CMakeLists.txt
@@ -139,6 +139,7 @@ add_plugin (crypto_gcrypt
 		ELEKTRA_PLUGIN_NAME=\"crypto_gcrypt\"
 		ELEKTRA_VARIANT=gcrypt
 		ELEKTRA_CRYPTO_API_GCRYPT
+	ADD_TEST
 )
 
 #
@@ -158,6 +159,7 @@ add_plugin (crypto_openssl
 		ELEKTRA_PLUGIN_NAME=\"crypto_openssl\"
 		ELEKTRA_VARIANT=openssl
 		ELEKTRA_CRYPTO_API_OPENSSL
+	ADD_TEST
 )
 
 #
@@ -177,13 +179,6 @@ add_plugin (crypto_botan
 		ELEKTRA_PLUGIN_NAME=\"crypto_botan\"
 		ELEKTRA_VARIANT=botan
 		ELEKTRA_CRYPTO_API_BOTAN
+	ADD_TEST
 )
 
-if (ADDTESTING_PHASE)
-	if (HAS_OPENSSL_4SURE OR HAS_GCRYPT_4SURE OR HAS_BOTAN_4SURE)
-		add_plugintest (crypto
-				MEMLEAK
-				LINK_PLUGIN "<no>")
-		set_property (TEST testmod_crypto PROPERTY LABELS memleak)
-	endif ()
-endif ()
